Add files via upload
Health Diagnosis one of the most advanced data science medical projects that I have done till now in my whole Data Science journey. which is an all-in-one medical solution. We know the future is all about AI so here is my idea of bringing 7 disease detections under one platform using the power of AI. Although these results are perfect yet an experiment we did, which can prove to be a revolution in the coming years.